fix: extract setup-ui.tsx, remove Docker refs from docs (Copilot review)
