+++ app/setup/_components/setup-flow.tsx
@@ -13,7 +13,8 @@ import {
 import { Label } from "@/components/ui/label";
 import { Loader2, Eye, EyeOff, Shield, AlertCircle, Store } from "lucide-react";
 import { useSiteSettings } from "@/hooks/useSiteSettings";
-import { EulaStep, SetupStepper, SetupHeader, SetupLogo } from "./eula-step";
+import { EulaStep } from "./eula-step";
+import { SetupStepper, SetupHeader, SetupLogo } from "./setup-ui";
 import type { LegalDocument } from "@/lib/legal-utils";
 
 interface PasswordValidation {
